requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
rich>=13.0.0
openai-whisper>=20231117
sqlalchemy>=2.0.0
duckduckgo-search>=5.0.0
